+++ lora/CHANGELOG.md
@@ -8,4 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
-- Move IQ inversion out of `RxConfig` and `TxConfig` into dedicated `set_iq_inversion` driver method. It appears that TX path IQ inversion defaults to on, which was breaking all examples.
+- move IQ inversion out of `RxConfig` and `TxConfig` into dedicated `set_iq_inversion` driver method.
+- move preamble set to `Sx127xLoraConfig`
+- make `optimize_rx_response` public
+
+### Removed
+- `RxConfig` struct
+- `config_rx` method